Restructure project files
